Return size_t from the desciption functions
This is a response to the compiler warning about the return
value being uninitialized in describe_row.  It seems appropriate,
since this value is passed as the 3rd argument to write, which expects
a size_t.  We are not using -1 as an error value, so there is no point
in returning it.

+++ smtx-test.h
@@ -22,9 +22,9 @@
  * access the tests have to the program.
  */
 
-extern unsigned describe_layout(char *, ptrdiff_t, unsigned);
-extern unsigned describe_row(char *, size_t, int);
-extern unsigned describe_state(char *desc, size_t siz);
+extern size_t describe_layout(char *, ptrdiff_t, unsigned);
+extern size_t describe_row(char *, size_t, int);
+extern size_t describe_state(char *desc, size_t siz);
 extern int smtx_main(int, char **);
 
 #define CTL(x) ((x) & 0x1f)
